% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config_file}
\alias{get_config_file}
\title{Get the template to be configured to process the files with
physical activity data}
\usage{
get_config_file()
}
\value{
A message informing on the success or the failure of the creation of the \code{config.csv} file.
}
\description{
The function creates a \code{config.csv} file at the root of the current working directory.
This file is needed to perform batch analysis. If there already is a file named \code{config.csv}
at the root of the current working directory, the file will not be created.
}
\examples{
\dontrun{
get_config_file()
}
}
