% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates_template.R
\name{get_dates_template}
\alias{get_dates_template}
\title{Get a template of the file where to indicate the dates (start and end) of the physical activity
measurement conducted in the participant(s)}
\usage{
get_dates_template()
}
\value{
A message informing on the success or the failure of the creation of the \code{DATES.csv} file.
}
\description{
This function creates a \code{DATES.csv} file in a \code{data} folder placed at the root
of the current working directory. This file is a template to be filled so that it
contains, for the participant(s) to analyse, the start and end dates of the measurement.
}
\examples{
\dontrun{
get_dates_template()
}
}
