% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demo_template.R
\name{get_demo_template}
\alias{get_demo_template}
\title{Get a template of the file where to indicate the demographic characteristics of the participant(s) to be analysed}
\usage{
get_demo_template()
}
\value{
A message informing on the success or the failure of the creation of the \code{DEMO.csv} file.
}
\description{
This function creates a \code{DEMO.csv} file in a \code{data} folder placed at the root
of  the current working directory. This file is a template to be filled so that
it contains, for the participant(s) to analyse, the following variables: id number,
sex, age (yr), and weight (kg).
}
\examples{
\dontrun{
get_demo_template()
}
}
