% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_agd.R
\name{process_agd}
\alias{process_agd}
\title{Process a single .agd file}
\usage{
process_agd(
  agd_file,
  config,
  demo,
  dates = NULL,
  id_config,
  content = c("option_1", "option_2", "option_3")
)
}
\arguments{
\item{agd_file}{A character string to set the path to the .agd file. The name
of the file shoud begin with the participant number id (e.g., "001").}

\item{config}{A dataframe obtained after importing the \code{config.csv} file needed
to allow the analysis process. Use the \code{\link{get_config_file}} function to
get the default \code{config.csv} file.}

\item{demo}{A dataframe containing the demographic variables needed to
compute physical activity metrics, that are:
\itemize{
\item \code{id}: participant inclusion number.
\item \code{age}: age of the participant.
\item \code{sex}: sex of the participant. Males could be indicated by
\code{male}, and females by \code{female}.
\item \code{weight}: weight of the participant in kg.
}
The names of these variables can be differently written in your own file
containing these variables and should be set in the \code{config.csv} file to
be filled to allow the analysis process. Use the \code{\link{get_demo_template}}
function to get a template of the file to be imported with the required demographic
characteristics.}

\item{dates}{A dataframe containing the start and end dates of the physical
activity measurement. Use the \code{\link{get_dates_template}} function to
get a template of the file to be imported with the required dates if wanted.
Default is \code{NULL}. If \code{NULL},  all the days of measurement will be analysed.
If an appropriate dataframe is provided, the analyses will be conducted
using the dates provided in the dataframe.}

\item{id_config}{A numeric value to indicate how many numbers are used to
identify the participant in the name of the .agd file (e.g., 2 for "01",
3 for "001", 4 for "0001", etc). The name of the file to be analysed has to
begin by the participant number written as in the mentionned examples above.}

\item{content}{A character value to indicate what results should be contained
in the list of the exported results:
\itemize{
\item \code{option_1} provides only the metrics summarised using valid days.
\item \code{option_2} provides, in addition to \code{option_1}, the results computed
for each day of measurement.
\item \code{option_3} provides, in addition to \code{option_2}, the initial dataset
with all the marks of the computations performed during the analysis process.
}}
}
\value{
The datasets required by the \code{content} argument of the function:
\itemize{
\item \code{all_metrics}: the metrics summarized using the valid days.
\item \code{results_by_day}: the results computed for each day of measurement
(\code{option_2} and \code{option_3} only).
\item \code{data_with_intensity_marks}: the initial dataset with all the
marks of the computations performed during the analysis process (\code{option_3}
only).
}
If the \code{content} argument is \code{option_1}, a dataframe is returned. Otherwise
a list is returned.
}
\description{
The function analyses a .agd file using the functions from the activAnalyzer
R package and using information relating to the demographic characteristics
of the participant, the dates (start and end) of the physical activity
measurement (optional), and the processing settings.
}
\examples{
FILE_PATH <- system.file("agd/001.agd", package = "activAnalyzer.batch")
CONFIG <- readr::read_csv2(system.file("config.csv", package = "activAnalyzer.batch"))
DEMO <- readr::read_csv2(system.file("DEMO.csv", package = "activAnalyzer.batch"))
DATES <- readr::read_csv2(system.file("DATES.csv", package = "activAnalyzer.batch"))
process_agd(
agd_file = FILE_PATH, 
config = CONFIG, 
demo = DEMO, 
dates = DATES, 
id_config = 3
)
}
