% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all_agd.R
\name{process_all_agd}
\alias{process_all_agd}
\title{Process all .agd files}
\usage{
process_all_agd(
  agd_dir,
  config_path,
  demo_path,
  dates_path = NULL,
  id_config,
  content = c("option_1", "option_2", "option_3"),
  out_dir = "./out",
  export_results = FALSE,
  cores = NULL
)
}
\arguments{
\item{agd_dir}{A character string to set the path to the directory where are all
the .agd files to be analysed. The names of the files shoud begin with the
participant number id (e.g., "001").}

\item{config_path}{A character string to set the path to the \code{config.csv} file to
be filled to allow the analysis process. Use the \code{\link{get_config_file}}
function to get the default \code{config.csv} file.}

\item{demo_path}{A character string to set the path to the .csv file that should
contain the demographic variables needed to compute physical activity metrics,
that are:
\itemize{
\item \code{id}: participant inclusion number.
\item \code{age}: age of the participant.
\item \code{sex}: sex of the participant. Males could be indicated by
\code{male}, and females by \code{female}.
\item \code{weight}: weight of the participant in kg.
}
The names of these variables can be differently written in your own file
containing these variables and should be set in the \code{config.csv} file to
be filled to allow the analysis process. Use the \code{\link{get_demo_template}}
function to get a template of the file to be imported with the required demographic
characteristics.}

\item{dates_path}{A character string to set the path to the .csv file that
contains the start and end dates of the physical activity measurement(s).
Use the \code{\link{get_dates_template}} function to get a template of the
file to be imported with the required dates if wanted. Default is \code{NULL}.
If \code{NULL},  all the days of measurement will be analysed. If the path to an
appropriate file is provided, the analyses will be conducted using
the dates provided in the file.}

\item{id_config}{A numeric value to indicate how many numbers are used to
identify the participant in the names of the .agd files (e.g., 2 for "01", 3 for "001",
4 for "0001", etc). The names of the files to be analysed have to
begin by the participant number written as in the mentionned examples above.}

\item{content}{A character value to indicate what results should be contained
in the list of the exported results:
\itemize{
\item \code{option_1} provides only the metrics summarised using valid days.
\item \code{option_2} provides, in addition to \code{option_1}, the results computed
for each day of measurement.
\item \code{option_3} provides, in addition to \code{option_2}, the initial dataset(s)
with all the marks of the computations performed during the analysis process.
}}

\item{out_dir}{A character string to set the path to the directory that will be used
to export the results if required.}

\item{export_results}{A logical value regarding the export of the results. The
results are exported as a dataframe (if \code{content} argument = "option_1")
or as a list object otherwise, embedded in a .rds file. The results exported are
the followings:
\itemize{
\item \code{option_1}: the metrics summarized using the valid days only.
\item \code{option_2}: the results from \code{option_1} plus the results computed for each
day of measurement.
\item \code{option_3}: the results from \code{option_2} plus the initial dataset(s) with all the
marks of the computations performed during the analysis process.
}}

\item{cores}{A positive integer representing the number of cores to use for parallel
processing (i.e., the number of processes to start). This value must be between 2 and
parallel::detectCores() - 1. Default is parallel::detectCores() - 1.}
}
\value{
The data required by the \code{content} argument of the function:
\itemize{
\item \code{all_metrics}: the metrics summarized using the valid days and collapsed
into a single dataframe.
\item \code{results_by_day}: the results computed for each day of measurement
collapsed into a single dataframe (\code{option_2} and \code{option_3} only).
\item \code{data_with_intensity_marks}: the initial dataset(s) with all the
marks of the computations performed during the analysis process collapsed
into a single dataframe (\code{option_3} only).
}
If the \code{content} argument is \code{option_1}, a dataframe is returned. Otherwise
a list of dataframes is returned.
}
\description{
The function analyses all the .agd files present in a given directory using
the functions from the activAnalyzer R package and using information
relating to the demographic characteristics of the participant(s),
the dates (start and end) of the physical activity measurement(s) (optional),
and the processing settings.
}
\examples{
DIR_AGD_PATH <- system.file("agd/", package = "activAnalyzer.batch")
DEMO_PATH <- system.file("DEMO.csv", package = "activAnalyzer.batch")
DATES_PATH <- system.file("DATES.csv", package = "activAnalyzer.batch")
CONFIG_PATH <- system.file("config.csv", package = "activAnalyzer.batch")
process_all_agd(
agd_dir = DIR_AGD_PATH, 
demo_path = DEMO_PATH, 
config_path = CONFIG_PATH,
dates_path = DATES_PATH, 
id_config = 3,
cores = 2
)
}
